replaced grunt build with npm scripts
