+++ .npmignore
@@ -1,3 +1,4 @@
+.nyc_output/
 coverage/
 test/
 .*
@@ -5,5 +6,4 @@ AUTHORS.md
 bower.json
 CHANGES.md
 CONTRIBUTING.md
-Gruntfile.js
 README.md
