+++ .nycrc
@@ -1,0 +1,10 @@
+{
+  "check-coverage": true,
+  "lines": 95,
+  "statements": 95,
+  "functions": 95,
+  "branches": 95,
+  "include": [ "src/**/*.js" ],
+  "reporter": [ "lcov" ],
+  "all": true
+}
