+++ bower.json
@@ -30,6 +30,7 @@
     "node"
   ],
   "ignore": [
+    ".nyc_output/",
     "coverage/",
     "src/",
     "test/",
@@ -37,7 +38,6 @@
     "AUTHORS.md",
     "CHANGES.md",
     "CONTRIBUTING.md",
-    "Gruntfile.js",
     "package.json",
     "README.md"
   ]
