+++ codecov.yml
@@ -1,1 +1,5 @@
 comment: off
+coverage:
+  precision: 2
+  range: "95...100"
+  round: down
